joints 1 to 3 measured and displaying in rviz

+++ sr_ronex_arm_transmissions/src/mapping/general_io/command_to_servo.cpp
@@ -179,8 +179,7 @@ namespace ronex
 	if      (effort >  100.0) effort =  100.0;
 	else if (effort < -100.0) effort = -100.0;
 
-	return ( (effort + 100) / 200) * (max_on_time_ - min_on_time_) + min_on_time_;
-
+	return (unsigned short int) ( ( (effort + 100.0) / 200.0 ) * double(max_on_time_ - min_on_time_) + double(min_on_time_) );
       }
     }
   }
